pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.120.0",
    "groq>=0.33.0",
    "mcp>=1.19.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
